py-app/pyproject.toml
[project]
name = "py-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "websockets>=12.0",
    "pydantic>=2.0",
    # OpenTelemetry packages
    "opentelemetry-distro>=0.59b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
    "opentelemetry-instrumentation-redis>=0.59b0",
    "redis>=6.4.0",
]
